feat: integrate logging and improve error handling
- Added a logger initialization in main.go to create a log file for the application.
- Updated various files to replace fmt.Println with structured logging using slog for better error tracking.
- Removed unused file writing functions and debug logs.
- Enhanced user feedback in the UI by logging errors instead of printing them directly to the console.
