+++ internal/ui/modalForeground.go
@@ -132,217 +132,6 @@ var debouncedSearch = Debounce(func(args ...interface{}) tea.Msg {
 	return rpc.RpcClient.Search(query)
 }, 300*time.Millisecond)
 
-func (m *Foreground) Update(message tea.Msg) (tea.Model, tea.Cmd) {
-	var cmds []tea.Cmd
-	switch msg := message.(type) {
-	case tea.WindowSizeMsg:
-		m.searchResultCombined = list.New([]list.Item{}, SearchDelegate{}, 10, 10)
-
-		m.searchResultCombined.Title = "Search User Result"
-		m.searchResultCombined.SetShowStatusBar(false)
-		m.searchResultCombined.SetShowFilter(false)
-		m.searchResultCombined.SetShowStatusBar(false)
-		m.windowWidth = msg.Width
-		m.windowHeight = msg.Height
-		m.focusedOn = SEARCH
-		input := textinput.New()
-		input.Placeholder = "Search..."
-		input.Prompt = "🔍 "
-		input.CharLimit = 256
-		m.input = input
-		m.input.Focus()
-	case tea.KeyMsg:
-		switch msg.String() {
-		case "tab":
-			if m.focusedOn == SEARCH {
-				m.focusedOn = LIST
-				m.input.Blur()
-			} else {
-				m.focusedOn = SEARCH
-				m.input.Focus()
-			}
-		case "enter":
-			if m.ModalMode == ModalModeForwardMessage {
-				if m.UsersList != nil {
-					selectedUser := m.UsersList.SelectedItem()
-
-					closeCommandCMD := func() tea.Msg {
-						return CloseOverlay{}
-					}
-
-					var from list.Item
-					if m.fromPeer != nil {
-						from = *m.fromPeer
-					}
-
-					if selectedUser != nil {
-						return m, tea.Batch(closeCommandCMD, func() tea.Msg {
-							return ForwardMsg{
-								msg:      m.Message,
-								reciever: &selectedUser,
-								fromPeer: &from,
-							}
-						})
-					}
-				}
-			}
-			if m.focusedOn == LIST {
-				selectedUser := m.searchResultCombined.SelectedItem()
-				if selectedUser != nil {
-					user, ok := selectedUser.(SearchResult)
-					if ok {
-						closeCommandCMD := func() tea.Msg {
-							return CloseOverlay{}
-						}
-
-						var channel *rpc.ChannelAndGroupInfo = nil
-						if user.ChannelOrUserType == CHANNEL {
-							for _, v := range m.SearchResultChannels {
-								if v.ChannelID == user.PeerID {
-									channel = &v
-								}
-							}
-						}
-
-						var u *rpc.UserInfo = nil
-						if user.ChannelOrUserType == USER {
-							for _, v := range m.searchResultUsers {
-								if v.PeerID == user.PeerID {
-									u = &v
-								}
-							}
-						}
-
-						var group *rpc.ChannelAndGroupInfo = nil
-						if user.ChannelOrUserType == GROUP {
-							for _, v := range m.SearchResultChannels {
-								if v.ChannelID == user.PeerID {
-									group = &v
-								}
-							}
-						}
-
-						cmd := func() tea.Msg {
-							if u != nil {
-								return SelectSearchedUserResult{
-									user:    u,
-									channel: nil,
-								}
-							}
-							if channel != nil {
-								return SelectSearchedUserResult{
-									channel: channel,
-									user:    nil,
-								}
-
-							}
-							if group != nil {
-								return SelectSearchedUserResult{
-									group: group,
-									user:  nil,
-								}
-							}
-							return nil
-						}
-						return m, tea.Batch(cmd, closeCommandCMD)
-					}
-				}
-			}
-		case "y", "Y":
-			if m.ModalMode == ModalModeDeleteMessage {
-				closeCommandCMD := func() tea.Msg {
-					return CloseOverlay{}
-				}
-
-				return m, tea.Batch(closeCommandCMD, func() tea.Msg {
-					return MessageDeletionConfrimResponseMsg{yes: true}
-				})
-			}
-		case "n", "N":
-			if m.ModalMode == ModalModeDeleteMessage {
-				closeCommandCMD := func() tea.Msg {
-					return CloseOverlay{}
-				}
-				return m, tea.Batch(closeCommandCMD, func() tea.Msg {
-					return MessageDeletionConfrimResponseMsg{yes: false}
-				})
-			}
-		}
-
-		if m.input.Focused() {
-			searchValue := m.input.Value()
-			if len(searchValue) >= 3 {
-				searchCmd := debouncedSearch(searchValue)
-				cmds = append(cmds, searchCmd)
-			}
-		}
-
-	case rpc.SearchUserMsg:
-		if msg.Err != nil {
-			//TODO:show error message here
-		} else {
-			// Wait for the message and return it
-			result := msg.Response.Result
-			var users []list.Item
-			for _, v := range result.Users {
-				users = append(users, SearchResult{
-					Name:              v.FirstName,
-					IsBot:             v.IsBot,
-					PeerID:            v.PeerID,
-					AccessHash:        v.AccessHash,
-					UnreadCount:       v.UnreadCount,
-					ChannelOrUserType: USER,
-				})
-			}
-
-			for _, v := range result.Channels {
-				var channelOrGroup ChannelOrUserType
-				if v.IsBroadcast {
-					channelOrGroup = CHANNEL
-				} else {
-					channelOrGroup = GROUP
-				}
-				users = append(users, SearchResult{
-					Name:              v.ChannelTitle,
-					IsBot:             false,
-					PeerID:            v.ChannelID,
-					AccessHash:        v.AccessHash,
-					UnreadCount:       v.UnreadCount,
-					ChannelOrUserType: channelOrGroup,
-				})
-			}
-			setTotalSearchResultUsers(msg, m)
-			m.searchResultCombined.SetItems(users)
-		}
-	case OpenModalMsg:
-		m.ModalMode = msg.ModalMode
-		m.Message = msg.Message
-		m.fromPeer = msg.FromPeer
-		m.UsersList = msg.UsersList
-		if msg.ModalMode == ModalModeSearch {
-			m.focusedOn = SEARCH
-		}
-	}
-
-	input, cmd := m.input.Update(message)
-	m.input = input
-
-	if m.UsersList != nil {
-		userList, userListCmd := m.UsersList.Update(message)
-		m.UsersList = &userList
-		cmds = append(cmds, userListCmd)
-	}
-
-	cmds = append(cmds, cmd)
-	if m.focusedOn == LIST {
-		users, userCmd := m.searchResultCombined.Update(message)
-		m.searchResultCombined = users
-		cmds = append(cmds, userCmd)
-	}
-
-	return m, tea.Batch(cmds...)
-}
-
 func setTotalSearchResultUsers(searchMsg rpc.SearchUserMsg, m *Foreground) {
 	m.searchResultUsers = searchMsg.Response.Result.Users
 	m.SearchResultChannels = searchMsg.Response.Result.Channels
