+++ js/package.json
@@ -21,9 +21,9 @@
     "url": "https://github.com/kumneger0/cligram"
   },
   "scripts": {
-	"dev": "nodemon --watch src --watch test --ext ts,json --exec 'bun run build'",
-	"build": "bun build src/index.ts --compile --outfile bin/cligram-js",
-	"build:esm": "tsdown src/index.ts --format esm",
+    "dev": "nodemon --watch src --watch test --ext ts,json --exec 'bun run build'",
+    "build": "bun build src/index.ts --compile --outfile bin/cligram-js",
+    "build:esm": "tsdown src/index.ts --format esm",
     "start": "bun src/index.ts",
     "test:": "vitest",
     "format": "prettier . --write",
@@ -33,45 +33,34 @@
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
-    "@tsconfig/node20": "^20.1.4",
     "@types/bun": "^1.2.2",
     "@typescript-eslint/eslint-plugin": "6.21.0",
     "@typescript-eslint/parser": "6.21.0",
-    "@typescript/native-preview": "^7.0.0-dev.20250527.1",
-    "esbuild": "0.21.3",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
-	"nodemon": "^3.1.10",
+    "nodemon": "^3.1.10",
     "prettier": "3.3.3",
     "tsdown": "^0.12.7",
     "typescript": "5.5.4",
     "vitest": "^3.0.7"
   },
   "dependencies": {
-    "@types/ink": "^2.0.3",
     "@types/node-notifier": "^8.0.5",
-    "@types/react": "^19.0.8",
-    "clipboardy": "^4.0.0",
     "date-fns": "^4.1.0",
     "debounce": "^2.2.0",
     "dotenv": "^16.4.7",
-    "flatted": "^3.3.3",
     "input": "^1.0.1",
     "kolorist": "^1.8.0",
     "lru-cache": "^11.1.0",
     "node-notifier": "^10.0.1",
     "telegram": "^2.26.22",
-    "term-size": "^4.0.0",
-    "terminal-image": "^3.1.1",
-    "zod": "^3.24.2",
-    "zustand": "^5.0.3"
+    "zod": "^3.24.2"
   },
   "pnpm": {
     "onlyBuiltDependencies": [
       "bufferutil",
       "core-js",
       "es5-ext",
-      "esbuild",
       "utf-8-validate"
     ]
   }
