--- pkg/example/example.go
@@ -1,11 +1,0 @@
-package example
-
-// Add returns the summary of two integers.
-func Add(a, b int) int {
-	return a + b
-}
-
-// Multiply returns the product of two integers.
-func Multiply(a, b int) int {
-	return a * b
-}
